+++ Changelog.md
@@ -21,8 +21,9 @@
 - Remove `.ip_string()` for `.ip.to_string()`
 - Add `Response.close()` for closeing a stream with no response
 - Completly Redo Middleware, Now it can modify Requests / Responses
-- Removed Server.middleware
+- Removed Server.middleware()
 - Export Internal Functions
+- Supply the Request to the Middleware Post Function
 
 # 0.2.2
 - Remove Debug Print Left in...